+++ devenv.nix
@@ -5,7 +5,9 @@
   # see https://github.com/CQCL/tket2/blob/main/devenv.nix
   packages = [
     pkgs.just
-    pkgs.llvmPackages_14.libllvm
+    pkgs.graphviz
+
+    # These are required for hugr-llvm to be able to link to llvm.
     pkgs.libffi
     pkgs.libxml2
   ]
@@ -16,30 +18,29 @@
     with pkgs.darwin.apple_sdk; [
       frameworks.CoreServices
       frameworks.CoreFoundation
+      pkgs.zlib
+      pkgs.ncurses
     ]
   );
 
   enterShell = ''
-    export PATH="$UV_PROJECT_ENVIRONMENT/bin:$PATH"
+    export LD_LIBRARY_PATH="${lib.makeLibraryPath [ pkgs.stdenv.cc.cc.lib ]}''${LD_LIBRARY_PATH:+:$LD_LIBRARY_PATH}";
   '';
 
   languages.python = {
     enable = true;
     uv = {
       enable = true;
-      sync = {
-        enable = true;
-        allExtras = true;
-      };
+      sync.enable = true;
     };
+    venv.enable = true;
   };
 
   env.LLVM_SYS_140_PREFIX = pkgs.llvmPackages_14.libllvm.dev;
 
-  env.LD_LIBRARY_PATH = "${lib.makeLibraryPath [ pkgs.stdenv.cc.cc.lib ]}";
-
   languages.rust = {
     enable = true;
     channel = "stable";
+    components = [ "rustc" "cargo" "clippy" "rustfmt" "rust-analyzer" ];
   };
 }