+++ guppylang/__init__.py
@@ -7,4 +7,4 @@ from guppylang.std.quantum import qubit
 
 # This is updated by our release-please workflow, triggered by this
 # annotation: x-release-please-version
-__version__ = "0.16.0"
+__version__ = "0.17.1"