+++ guppylang/checker/linearity_checker.py
@@ -43,6 +43,7 @@ from guppylang.definition.value import CallableDef
 from guppylang.error import GuppyError, GuppyTypeError
 from guppylang.nodes import (
     AnyCall,
+    BarrierExpr,
     CheckedNestedFunctionDef,
     DesugaredArrayComp,
     DesugaredGenerator,
@@ -413,6 +414,10 @@ class BBLinearityChecker(ast.NodeVisitor):
         self.visit(node.item_expr)
         self.scope.assign(node.item)
         self.visit(node.getitem_expr)
+
+    def visit_BarrierExpr(self, node: BarrierExpr) -> None:
+        self._visit_call_args(node.func_ty, node)
+        self._reassign_inout_args(node.func_ty, node)
 
     def visit_Expr(self, node: ast.Expr) -> None:
         # An expression statement where the return value is discarded